+++ src-api/app/main.py
@@ -29,7 +29,6 @@ client.connect()
 db = Database()
 
 users_db = db.get_users()
-print(users_db['jerlends@tuta.io'].exists())
 
 
 app = FastAPI()
@@ -37,7 +36,12 @@ app = FastAPI()
 
 @app.post("/account/create")
 async def create_account(form_data: NewUser):
+    """
+    Creates an account where email is the unique field
 
+    :param form_data:
+    :return:
+    """
     if form_data.password_one != form_data.password_two:
         return {'error': 'Passwords do not match.'}
     password = form_data.password_one
@@ -78,7 +82,6 @@ async def create_account(form_data: NewUser):
 async def login_for_access_token(form_data: OAuth2PasswordRequestForm = Depends()):
     # username == _id == email
     user = authenticate_user(form_data.username, form_data.password)
-    print(user)
     if not user:
         raise HTTPException(
             status_code=status.HTTP_401_UNAUTHORIZED,
@@ -86,9 +89,8 @@ async def login_for_access_token(form_data: OAuth2PasswordRequestForm = Depends(
             headers={"WWW-Authenticate": "Bearer"},
         )
     access_token_expires = timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
-    print(user.username)
     access_token = create_access_token(
-        data={"sub": user.username}, expires_delta=access_token_expires
+        data={"sub": user._id}, expires_delta=access_token_expires
     )
     return {"access_token": access_token, "token_type": "bearer"}
 
@@ -98,9 +100,9 @@ async def read_users_me(current_user: User = Depends(get_current_active_user)):
     return current_user
 
 
-@app.get("/users/account}/")
+@app.get("/users/account/")
 async def read_own_items(current_user: User = Depends(get_current_active_user)):
-    return [{"item_id": "Foo", "owner": current_user.username}]
+    return [{"owner": current_user._id}]
 
 
 @app.on_event("startup")